DAY 26

+++ server/src/main/java/com/edutech/progressive/service/impl/BillingServiceImpl.java
@@ -2,12 +2,12 @@ package com.edutech.progressive.service.impl;
 
 import java.util.List;
 
+import com.edutech.progressive.entity.Billing;
+import com.edutech.progressive.service.BillingService;
+import com.edutech.progressive.repository.BillingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.edutech.progressive.entity.Billing;
-import com.edutech.progressive.repository.BillingRepository;
-import com.edutech.progressive.service.BillingService;
 
 @Service
 public class BillingServiceImpl implements BillingService {
@@ -16,29 +16,31 @@ public class BillingServiceImpl implements BillingService {
     private BillingRepository billingRepository;
 
     @Override
-    public List<Billing> getAllBills() {
+    public List<Billing> getAllBills() throws Exception {
         return billingRepository.findAll();
     }
 
     @Override
-    public Billing getBillById(int billingId) {
-        return billingRepository.findById(billingId).get();
+    public Billing getBillById(int billingId) throws Exception {
+        return billingRepository.findById(billingId)
+                .orElseThrow(() -> new Exception("Billing record not found with ID: " + billingId));
     }
 
     @Override
-    public Integer createBill(Billing billing) {
+    public Integer createBill(Billing billing) throws Exception {
         return billingRepository.save(billing).getBillingId();
     }
 
     @Override
-    public void deleteBill(int billingId) {
+    public void deleteBill(int billingId) throws Exception {
+        if (!billingRepository.existsById(billingId)) {
+            throw new Exception("Billing record not found with ID: " + billingId);
+        }
         billingRepository.deleteById(billingId);
     }
 
     @Override
-    public List<Billing> getBillsByPatientId(int patientId){
+    public List<Billing> getBillsByPatientId(int patientId) throws Exception {
         return billingRepository.findAllByPatient_PatientId(patientId);
     }
-
-
 }
